+++ sky-pojo/src/main/java/com/sky/dto/ShoppingCartDTO.java
@@ -6,8 +6,8 @@ import java.io.Serializable;
 @Data
 public class ShoppingCartDTO implements Serializable {
 
-    private Long dishId;
-    private Long setmealId;
+    private Integer dishId;
+    private Integer setmealId;
     private String dishFlavor;
 
 }
