+++ sky-pojo/src/main/java/com/sky/entity/Dish.java
@@ -19,13 +19,13 @@ public class Dish implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
 
     //菜品名称
     private String name;
 
     //菜品分类id
-    private Long categoryId;
+    private Integer categoryId;
 
     //菜品价格
     private BigDecimal price;
@@ -43,8 +43,8 @@ public class Dish implements Serializable {
 
     private LocalDateTime updateTime;
 
-    private Long createUser;
+    private Integer createUser;
 
-    private Long updateUser;
+    private Integer updateUser;
 
 }
