+++ sky-pojo/src/main/java/com/sky/entity/ShoppingCart.java
@@ -20,19 +20,19 @@ public class ShoppingCart implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
 
     //名称
     private String name;
 
     //用户id
-    private Long userId;
+    private Integer userId;
 
     //菜品id
-    private Long dishId;
+    private Integer dishId;
 
     //套餐id
-    private Long setmealId;
+    private Integer setmealId;
 
     //口味
     private String dishFlavor;
