+++ sky-pojo/src/main/java/com/sky/vo/DishVO.java
@@ -17,11 +17,11 @@ import java.util.List;
 @AllArgsConstructor
 public class DishVO implements Serializable {
 
-    private Long id;
+    private Integer id;
     //菜品名称
     private String name;
     //菜品分类id
-    private Long categoryId;
+    private Integer categoryId;
     //菜品价格
     private BigDecimal price;
     //图片
